Added in individual frame value indexing.
Saves an interesting value per frame to the file "output.txt", so it currently only works for one video at a time.
Prints the number of frames in the video to the first line of "output.txt"
Every subsequent line is the interesting factor of the frames in order starting with whatever frame was targeted at the command line.
Will make this file more descriptive and functional soon.